% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_BPREM.R
\name{Bayes_BPREM}
\alias{Bayes_BPREM}
\title{Bayesian Bivariate Piecewise Random Effects Model (BPREM)}
\usage{
Bayes_BPREM(
  data,
  id_var,
  time_var,
  y1_var,
  y2_var,
  iters_adapt = 5000,
  iters_burn_in = 1e+05,
  iters_sampling = 50000,
  thin = 15,
  save_full_chains = FALSE,
  save_conv_chains = FALSE
)
}
\arguments{
\item{save_conv_chains}{}
}
\description{
A short description...
}
\section{Details}{

}

\references{
reference
}
\author{
Corissa T. Rohloff
}
