% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_CREM.R
\name{Bayes_CREM}
\alias{Bayes_CREM}
\title{Bayesian Crossed Random Effects Model (CREM)}
\usage{
Bayes_CREM(
  data,
  ind_id_var,
  cross_id_var,
  time_var,
  y_var,
  form = "linear",
  fixed_effects = NULL,
  iters_adapt = 5000,
  iters_burn_in = 50000,
  iters_sampling = 50000,
  thin = 15,
  save_full_chains = FALSE,
  save_conv_chains = FALSE
)
}
\arguments{
\item{save_conv_chains}{}
}
\description{
A short description...
}
\section{Details}{

}

\references{
reference
}
\author{
Corissa T. Rohloff
}
