% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_PREM.R
\name{Bayes_PREM}
\alias{Bayes_PREM}
\title{Bayesian Piecewise Random Effects Model (PREM) + Extensions}
\usage{
Bayes_PREM(
  data,
  id_var,
  time_var,
  y_var,
  n_class = 1,
  max_cp = 2,
  class_predictive_vars = NULL,
  outcome_predictive_vars = NULL,
  scale_prior = "uniform",
  alpha = 1,
  cp_prior = "binomial",
  binom_prob = 0.5,
  iters_adapt = 1000,
  iters_burn_in = 20000,
  iters_sampling = 30000,
  thin = 15,
  save_full_chains = FALSE,
  save_conv_chains = FALSE
)
}
\arguments{
\item{save_conv_chains}{}
}
\description{
A short description...
}
\section{Details}{

}

\references{
reference
}
\author{
Corissa T. Rohloff
}
