% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClassProb.R
\name{getClassProb}
\alias{getClassProb}
\alias{getClassProb.PREM}
\alias{print.getClassProb.PREM}
\title{Extract class probabilities}
\usage{
getClassProb(object, ...)

\method{getClassProb}{PREM}(object, ...)

\method{print}{getClassProb.PREM}(object, ...)
}
\arguments{
\item{object}{An object of class "PREM".}

\item{...}{Additional arguments.}
}
\value{
Returns a list of the class probabilities.
}
\description{
Extracts the class probabilities from a fitted model of class "PREM".
}
\examples{
# load fitted model results
data(results_prem)
# get class probabilities
getClassProb(results_prem)

}
\seealso{
\link{Bayes_PREM}
}
\author{
Corissa T. Rohloff
}
