% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFitted.R
\name{getFitted}
\alias{getFitted}
\alias{getFitted.BEND}
\alias{print.getFitted}
\title{Extract fitted values}
\usage{
getFitted(x, ...)

\method{getFitted}{BEND}(x, ...)

\method{print}{getFitted}(x, ...)
}
\arguments{
\item{x}{An object of class "BPREM", "CREM", or "PREM".}

\item{...}{Additional arguments.}
}
\value{
Returns a data frame of the fitted values for each individual and timepoint.
}
\description{
Extracts the individual fitted values from a fitted model of class "BPREM", "CREM", or "PREM".
}
\examples{
# load fitted model results
data(results_prem)
# get fitted values
getFitted(results_prem)

}
\author{
Corissa T. Rohloff
}
