% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKProb.R
\name{getKProb}
\alias{getKProb}
\alias{getKProb.PREM}
\alias{print.getKProb.PREM}
\title{Extract changepoint probabilities}
\usage{
getKProb(object, ...)

\method{getKProb}{PREM}(object, ...)

\method{print}{getKProb.PREM}(object, ...)
}
\arguments{
\item{object}{An object of class "PREM".}

\item{...}{Additional arguments.}
}
\value{
Returns a list of the changepoint probabilities for each class .
}
\description{
Extracts the K (number of changepoints) probabilities from a fitted model of class "PREM".
}
\examples{
# load fitted model results
data(results_prem)
# get changepoint probabilities
getKProb(results_prem)

}
\seealso{
\link{Bayes_PREM}
}
\author{
Corissa T. Rohloff
}
