% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.CREM}
\alias{plot.CREM}
\title{Plot the results of a crossed random effects model (CREM)}
\usage{
\method{plot}{CREM}(
  x,
  xlab = "X",
  ylab = "Y",
  colors = NULL,
  mean_colors = NULL,
  legend_pos = "topright",
  ...
)
}
\arguments{
\item{x}{An object of class "CREM" (returned by \code{Bayes_CREM(...)}).}

\item{xlab}{X-axis label for the generated plot.}

\item{ylab}{Y-axis label for the generated plot.}

\item{colors}{Color for observed trajectories (optional). Default is "grey".}

\item{mean_colors}{Colors for the trajectory defined by the mean parameters for each outcome (optional). Default is "black".}

\item{legend_pos}{(optional) Option to change legend position (default = "topright").}

\item{...}{(optional) Other parameters to pass to the \code{plot()} function.}
}
\value{
No return value.
}
\description{
Provides a fitted plot of a CREM model, as returned by \code{Bayes_CREM()}.
}
\examples{
# load fitted model results
data(results_pcrem)
# plot fitted results
plot(results_pcrem)

}
\author{
Corissa T. Rohloff
}
