% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEND.R, R/data.R
\docType{data}
\name{results_bprem}
\alias{results_bprem}
\title{Fitted results for a BPREM}
\format{
A list (an object of class \code{BPREM}) with fitted model results.

\subsection{\code{results_bprem}}{

A list object with elements:
\describe{
\item{Call}{Function arguments}
\item{Sample_Size}{Data set sample size}
\item{Data}{Data used for model fitting}
\item{Convergence}{Convergence status}
\item{Model_Fit}{Model fit information}
\item{Fitted_Values}{Fitted outcome values}
\item{Parameter_Estimates}{Parameter estimates}
\item{Random_Coefficients}{Random coefficient estimates}
\item{Run_Time}{Model run time}
}
}
}
\usage{
data(results_bprem)

results_bprem
}
\description{
Simulated data for a bivariate piecewise random effects model (BPREM) using \code{SimData_BPREM}. Included to demonstrate the use of \code{summary.BPREM()}.

Results from fitting a bivariate piecewise random effects model (BPREM) to \code{SimData_BPREM} (returned by \code{Bayes_BPREM(...)}).
}
\keyword{datasets}
