% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEND.R, R/data.R
\docType{data}
\name{results_pcrem}
\alias{results_pcrem}
\title{Fitted results for a PCREM}
\format{
A list (an object of class \code{CREM}) with fitted model results.

\subsection{\code{results_pcrem}}{

A list object with elements:
\describe{
\item{Call}{Function arguments}
\item{Sample_Size}{Data set sample size}
\item{Data}{Data used for model fitting}
\item{Convergence}{Convergence status}
\item{Model_Fit}{Model fit information}
\item{Fitted_Values}{Fitted outcome values}
\item{Parameter_Estimates}{Parameter estimates}
\item{Random_Coefficients}{Random coefficient estimates}
\item{Run_Time}{Model run time}
}
}
}
\usage{
data(results_pcrem)

results_pcrem
}
\description{
Simulated data for a piecewise crossed random effects model (PCREM) using \code{SimData_CREM}. Included to demonstrate the use of \code{summary.CREM()}.

Results from fitting a piecewise crossed random effects model (PCREM) to \code{SimData_PCREM} (returned by \code{Bayes_CREM(...)}).
}
\keyword{datasets}
