% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEND.R, R/data.R
\docType{data}
\name{results_prem}
\alias{results_prem}
\title{Fitted results for a PREM}
\format{
A list (an object of class \code{PREM}) with fitted model results.

\subsection{\code{results_prem}}{

A list object with elements:
\describe{
\item{Call}{Function arguments}
\item{Sample_Size}{Data set sample size}
\item{Data}{Data used for model fitting}
\item{Convergence}{Convergence status}
\item{Model_Fit}{Model fit information}
\item{Fitted_Values}{Fitted outcome values}
\item{Parameter_Estimates}{Parameter estimates}
\item{Random_Coefficients}{Random coefficient estimates}
\item{Run_Time}{Model run time}
}
}
}
\usage{
data(results_prem)

results_prem
}
\description{
Fitted results for a piecewise random effects model (PREM) using \code{SimData_PREM}. Included to demonstrate the use of \code{plot_BEND()} and \code{summary.PREM()}.

Results from fitting a piecewise random effects model (PREM) to \code{SimData_PREM} (returned by \code{Bayes_PREM(...)}).
}
\keyword{datasets}
