% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEND.R, R/data.R
\docType{data}
\name{SimData_BPREM}
\alias{SimData_BPREM}
\title{Simulated data for a BPREM}
\format{
A data frame with 210 rows and 4 variables.

\subsection{\code{SimData_BPREM}}{

A data frame with 210 rows and 4 columns:
\describe{
\item{id}{Individual ids}
\item{time}{Time}
\item{y1}{Outcome 1}
\item{y2}{Outcome 2}
}
}
}
\usage{
data(SimData_BPREM)

SimData_BPREM
}
\description{
Simulated data for a bivariate piecewise random effects model (BPREM) with 7 timepoints collected on 30 individuals.

A simulated dataset for the bivariate piecewise random effects model (BPREM).
}
\details{
\itemize{
\item \code{id} ID for each individual.
\item \code{time} Timepoints for each individual.
\item \code{y1} Outcome 1.
\item \code{y2} Outcome 2.
}
}
\keyword{datasets}
