% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEND.R
\docType{data}
\name{SimData_CIPREMM}
\alias{SimData_CIPREMM}
\title{Simulated data for a CI-PREMM}
\format{
A data frame with 540 rows and 6 variables
}
\usage{
data(SimData_CIPREMM)
}
\description{
Simulated data for a covariate-influenced piecewise random effects mixture model (CI-PREMM) with 18 timepoints collected on 30 individuals. The variables are as follows:
}
\details{
\itemize{
\item id - id for each individual
\item time - timepoints for each individual
\item y - outcome
\item class_pred_1 - first class predictive covariate (time-invariant)
\item class_pred_2 - second class predictive covariate (time-invariant)
\item outcome_pred_1 - outcome predictive covariate (time-varying)
}
}
\keyword{datasets}
