% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEND.R, R/data.R
\docType{data}
\name{SimData_PCREM}
\alias{SimData_PCREM}
\title{Simulated data for a PCREM}
\format{
A data frame with 210 rows and 4 variables.

\subsection{\code{SimData_PCREM}}{

A data frame with 210 rows and 4 columns:
\describe{
\item{id}{Individual ids}
\item{teacherid}{Teacher ids}
\item{time}{Time}
\item{y}{Outcome}
}
}
}
\usage{
data(SimData_PCREM)

SimData_PCREM
}
\description{
Simulated data for a piecewise crossed random effects model (PCREM) with 7 timepoints collected on 30 individuals.

A simulated dataset for the piecewise crossed random effects model (PCREM).
}
\details{
\itemize{
\item \code{id} ID for each individual.
\item \code{teacherid} ID for each teacher.
\item \code{time} Timepoints for each individual.
\item \code{y} Outcome.
}
}
\keyword{datasets}
