% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEND.R, R/data.R
\docType{data}
\name{SimData_PREM}
\alias{SimData_PREM}
\title{Simulated data for a PREM + Extensions}
\format{
A data frame with 540 rows and 6 variables.

\subsection{\code{SimData_PREM}}{

A data frame with 540 rows and 6 columns:
\describe{
\item{id}{Individual ids}
\item{time}{Time}
\item{y}{Outcome}
\item{class_pred_1}{First Class-Predictive Covariate}
\item{class_pred_2}{Second Class-Predictive Covariate}
\item{outcome_pred_1}{First Outcome-Predictive Covariate}
}
}
}
\usage{
data(SimData_PREM)

SimData_PREM
}
\description{
Simulated data for a piecewise random effects model (PREM) and useful extensions (CI-PREM, PREMM, CI-PREMM) with 18 timepoints collected on 30 individuals.

A simulated dataset for the piecewise random effects model (PREM).
}
\details{
\itemize{
\item \code{id} ID for each individual.
\item \code{time} Timepoints for each individual.
\item \code{y} Outcome.
\item \code{class_pred_1} First class predictive covariate (time-invariant).
\item \code{class_pred_2} Second class predictive covariate (time-invariant).
\item \code{outcome_pred_1} Outcome predictive covariate (time-varying).
}
}
\keyword{datasets}
