% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClassProb.R
\name{getClassProb}
\alias{getClassProb}
\alias{getClassProb.PREM}
\alias{print.getClassProb.PREM}
\title{Extract class probabilities}
\usage{
getClassProb(x, ...)

\method{getClassProb}{PREM}(x, ...)

\method{print}{getClassProb.PREM}(x, ...)
}
\arguments{
\item{x}{An object of class "PREM".}

\item{...}{Additional arguments.}
}
\value{
Returns a list of the class probabilities.
}
\description{
Extracts the class probabilities from a fitted model of class "PREM".
}
\examples{
# load fitted model results
data(results_prem)
# get class probabilities
getClassProb(results_prem)

}
\author{
Corissa T. Rohloff
}
