% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoef.R
\name{getCoef}
\alias{getCoef}
\alias{getCoef.BPREM}
\alias{getCoef.CREM}
\alias{getCoef.PREM}
\alias{print.getCoef}
\title{Extract random coefficients}
\usage{
getCoef(object, ...)

\method{getCoef}{BPREM}(object, ...)

\method{getCoef}{CREM}(object, ...)

\method{getCoef}{PREM}(object, ...)

\method{print}{getCoef}(object, ...)
}
\arguments{
\item{object}{An object of class "BPREM", "CREM", or "PREM".}

\item{...}{Additional arguments.}
}
\value{
Returns a data frame of the random coefficients for each individual/group.
}
\description{
Extracts the random coefficients from a fitted model of class "BPREM", "CREM", or "PREM".
}
\examples{
# load fitted model results
data(results_prem)
# get random coefficients
getCoef(results_prem)

}
\seealso{
\link{Bayes_PREM, Bayes_CREM, Bayes_PREM}
}
\author{
Corissa T. Rohloff
}
