% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFixEf.R
\name{getFixEf}
\alias{getFixEf}
\alias{getFixEf.BPREM}
\alias{getFixEf.CREM}
\alias{getFixEf.PREM}
\alias{print.getFixEf}
\title{Extract fixed effects parameter estimates}
\usage{
getFixEf(object, ...)

\method{getFixEf}{BPREM}(object, ...)

\method{getFixEf}{CREM}(object, ...)

\method{getFixEf}{PREM}(object, ...)

\method{print}{getFixEf}(object, ...)
}
\arguments{
\item{object}{An object of class "BPREM", "CREM", or "PREM".}

\item{...}{Additional arguments.}
}
\value{
Returns a vector or dataframe of fixed effects parameter values.
}
\description{
Extracts the fixed effects parameter estimates from a fitted model of class "BPREM", "CREM", or "PREM".
}
\examples{
# load fitted model results
data(results_bprem)
# get fixed effects
getFixEf(results_bprem)

}
\seealso{
\link{Bayes_PREM, Bayes_CREM, Bayes_PREM}
}
\author{
Corissa T. Rohloff
}
