% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFixEf.R
\name{getFixEf}
\alias{getFixEf}
\alias{getFixEf.BPREM}
\alias{getFixEf.CREM}
\alias{getFixEf.PREM}
\alias{print.getFixEf}
\title{Extract fixed effects parameter estimates}
\usage{
getFixEf(x, ...)

\method{getFixEf}{BPREM}(x, ...)

\method{getFixEf}{CREM}(x, ...)

\method{getFixEf}{PREM}(x, ...)

\method{print}{getFixEf}(x, ...)
}
\arguments{
\item{x}{An object of class "BPREM", "CREM", or "PREM".}

\item{...}{Additional arguments.}
}
\value{
Returns a vector or data frame of fixed effects parameter values.
}
\description{
Extracts the fixed effects parameter estimates from a fitted model of class "BPREM", "CREM", or "PREM".
}
\examples{
# load fitted model results
data(results_bprem)
# get fixed effects
getFixEf(results_bprem)

}
\author{
Corissa T. Rohloff
}
