% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelFit.R
\name{getModelFit}
\alias{getModelFit}
\alias{getModelFit.BPREM}
\alias{getModelFit.CREM}
\alias{getModelFit.PREM}
\alias{print.getModelFit}
\title{Extract model fit}
\usage{
getModelFit(object, ...)

\method{getModelFit}{BPREM}(object, ...)

\method{getModelFit}{CREM}(object, ...)

\method{getModelFit}{PREM}(object, ...)

\method{print}{getModelFit}(object, ...)
}
\arguments{
\item{object}{An object of class "BPREM", "CREM", or "PREM".}

\item{...}{Additional arguments.}
}
\value{
Returns a vector of the model fit information (deviance, pD, DIC).
}
\description{
Extracts the model fit information from a fitted model of class "BPREM", "CREM", or "PREM".
}
\examples{
# load fitted model results
data(results_prem)
# get model fit
getModelFit(results_prem)

}
\seealso{
\link{Bayes_PREM, Bayes_CREM, Bayes_PREM}
}
\author{
Corissa T. Rohloff
}
