% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRanEf.R
\name{getRanEf}
\alias{getRanEf}
\alias{getRanEf.BPREM}
\alias{getRanEf.CREM}
\alias{print.getRanEf}
\title{Extract random effects}
\usage{
getRanEf(object, ...)

\method{getRanEf}{BPREM}(object, ...)

\method{getRanEf}{CREM}(object, ...)

\method{print}{getRanEf}(object, ...)
}
\arguments{
\item{object}{An object of class "BPREM" or "CREM".}

\item{...}{Additional arguments.}
}
\value{
Returns a list of the random effects for each individual/group.
}
\description{
Extracts the random effects from a fitted model of class "BPREM" or "CREM".
}
\examples{
# load fitted model results
data(results_bprem)
# get random effects
getRanEf(results_bprem)

}
\seealso{
\link{Bayes_BPREM, Bayes_CREM}
}
\author{
Corissa T. Rohloff
}
