% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarCov.R
\name{getVarCov}
\alias{getVarCov}
\alias{getVarCov.BPREM}
\alias{getVarCov.CREM}
\alias{getVarCov.PREM}
\alias{print.getVarCov.BPREM}
\alias{print.getVarCov.CREM}
\alias{print.getVarCov.PREM}
\title{Extract random effects variance-covariance matrix}
\usage{
getVarCov(object, ...)

\method{getVarCov}{BPREM}(object, ...)

\method{getVarCov}{CREM}(object, ...)

\method{getVarCov}{PREM}(object, ...)

\method{print}{getVarCov.BPREM}(object, ...)

\method{print}{getVarCov.CREM}(object, ...)

\method{print}{getVarCov.PREM}(object, ...)
}
\arguments{
\item{object}{An object of class "BPREM", "CREM", or "PREM".}

\item{...}{Additional arguments.}
}
\value{
Returns a list of the random effects variance-covariance matrices.
}
\description{
Extracts the random effects variance-covariance matrix from a fitted model of class "BPREM", "CREM", or "PREM".
}
\examples{
# load fitted model results
data(results_prem)
# get random effects variance-covariance matrices
getVarCov(results_prem)

}
\seealso{
\link{Bayes_PREM, Bayes_CREM, Bayes_PREM}
}
\author{
Corissa T. Rohloff
}
