% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BPREM.R
\name{plot.BPREM}
\alias{plot.BPREM}
\title{Plot the results of a bivariate piecewise random effects model (BPREM)}
\usage{
\method{plot}{BPREM}(
  object,
  xlab = "X",
  ylab = "Y",
  colors = NULL,
  mean_colors = NULL,
  legend_pos = "topright",
  ...
)
}
\arguments{
\item{object}{An object of class "BPREM" (returned by \code{Bayes_BPREM(...)}).}

\item{xlab}{X-axis label for the generated plot.}

\item{ylab}{Y-axis label for the generated plot.}

\item{colors}{Colors for each class outcome. By default, up to 2 colors are provided in the following order: "blue" (outcome 1), "red" (outcome 2).}

\item{mean_colors}{Colors for the trajectory defined by the mean parameters for each outcome. By default, up to 5 colors are provided in the following order: "darkblue" (outcome 1), "darkred" (outcome 2).}

\item{legend_pos}{(optional) Option to change legend position (default = "topright").}

\item{...}{(optional) Other parameters to pass to the \code{plot()} function.}
}
\value{
No return value.
}
\description{
Provides a fitted plot of a BPREM model, as returned by \code{Bayes_BPREM()}.
}
\examples{
# load fitted model results
data(results_bprem)
# plot fitted results
plot(results_bprem)

}
\author{
Corissa T. Rohloff
}
