% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PREM.R
\name{plot.PREM}
\alias{plot.PREM}
\title{Plot the results of a piecewise random effects model (PREM)}
\usage{
\method{plot}{PREM}(
  object,
  xlab = "X",
  ylab = "Y",
  colors = NULL,
  mean_colors = NULL,
  legend_pos = "topright",
  ...
)
}
\arguments{
\item{object}{An object of class "PREM" (returned by \code{Bayes_PREM(...)}).}

\item{xlab}{X-axis label for the generated plot.}

\item{ylab}{Y-axis label for the generated plot.}

\item{colors}{Colors for each class (\code{PREMM} or \code{CI-PREMM}). By default, up to 5 colors are provided in the following order: "blue" (class 1), "red" (class 2), "green" (class 3), "gold" (class 4), "gray" (class 5).}

\item{mean_colors}{Colors for the trajectory defined by the mean parameters for each class (\code{PREMM} or \code{CI-PREMM}). By default, up to 5 colors are provided in the following order: "darkblue" (class 1), "darkred" (class 2), "darkgreen" (class 3), "gold4" (class 4), "darkgray" (class 5).}

\item{legend_pos}{(optional) Option to change legend position (default = "topright").}

\item{...}{(optional) Other parameters to pass to the \code{plot()} function.}
}
\value{
No return value.
}
\description{
Provides a fitted plot of a PREM model, as returned by \code{Bayes_PREM()}.
}
\examples{
# load fitted model results
data(results_prem)
# plot fitted results
plot(results_prem)

}
\author{
Corissa T. Rohloff
}
