% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_BPREM.R
\name{print.BPREM}
\alias{print.BPREM}
\title{Print the results of a bivariate piecewise random effects model (BPREM)}
\usage{
\method{print}{BPREM}(x, ...)
}
\arguments{
\item{x}{An object of class "BPREM" (returned by \code{Bayes_BPREM(...)}).}

\item{...}{Additional arguments.}
}
\value{
Returns a list of key parameter estimates.
}
\description{
Provides a summary of a BPREM model, as returned by \code{Bayes_BPREM()}.
}
\examples{
# load fitted model results
data(results_bprem)
# print results
print(results_bprem)

}
\author{
Corissa T. Rohloff
}
