% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_CREM.R
\name{print.CREM}
\alias{print.CREM}
\title{Print the results of a crossed random effects model (CREM)}
\usage{
\method{print}{CREM}(x, ...)
}
\arguments{
\item{x}{An object of class "CREM" (returned by \code{Bayes_CREM(...)}).}

\item{...}{Additional arguments.}
}
\value{
Returns a list of key parameter estimates.
}
\description{
Provides a summary of a CREM model, as returned by \code{Bayes_CREM()}.
}
\examples{
# load fitted model results
data(results_pcrem)
# print results
print(results_pcrem)

}
\author{
Corissa T. Rohloff
}
