% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_PREM.R
\name{print.PREM}
\alias{print.PREM}
\title{Print the results of a piecewise random effects model (PREM)}
\usage{
\method{print}{PREM}(x, ...)
}
\arguments{
\item{x}{An object of class "PREM" (returned by \code{Bayes_PREM(...)}).}

\item{...}{Additional arguments.}
}
\value{
Returns a list of key parameter estimates.
}
\description{
Provides a summary of a PREM model, as returned by \code{Bayes_PREM()}.
}
\examples{
# load fitted model results
data(results_prem)
# print results
print(results_prem)

}
\author{
Corissa T. Rohloff
}
