% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEND.R
\docType{data}
\name{results_pcrem}
\alias{results_pcrem}
\title{Fitted results for a PCREM}
\format{
A list (an object of class \code{CREM}) with fitted model results.
}
\usage{
data(results_pcrem)
}
\description{
Simulated data for a piecewise crossed random effects model (PCREM) using \code{SimData_CREM}. Included to demonstrate the use of \code{summary.CREM()}.
}
\keyword{datasets}
