% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEND.R
\docType{data}
\name{SimData_BPREM}
\alias{SimData_BPREM}
\title{Simulated data for a BPREM}
\format{
A data frame with 210 rows and 4 variables
}
\usage{
data(SimData_BPREM)
}
\description{
Simulated data for a bivariate piecewise random effects model (BPREM) with 7 timepoints collected on 30 individuals. The variables are as follows:
}
\details{
\itemize{
\item id - id for each individual
\item time - timepoints for each individual
\item y1 - outcome 1
\item y2 - outcome 2
}
}
\keyword{datasets}
