% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEND.R
\docType{data}
\name{SimData_CIPREMM}
\alias{SimData_CIPREMM}
\title{Simulated data for a CI-PREMM}
\format{
A data frame with 540 rows and 6 variables.
}
\usage{
data(SimData_CIPREMM)
}
\description{
Simulated data for a covariate-influenced piecewise random effects mixture model (CI-PREMM) with 18 timepoints collected on 30 individuals.
}
\details{
\itemize{
\item \code{id} ID for each individual.
\item \code{time} Timepoints for each individual.
\item \code{y} Outcome.
\item \code{class_pred_1} First class predictive covariate (time-invariant).
\item \code{class_pred_2} Second class predictive covariate (time-invariant).
\item \code{outcome_pred_1} Outcome predictive covariate (time-varying).
}
}
\keyword{datasets}
