% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEND.R
\docType{data}
\name{SimData_PCREM}
\alias{SimData_PCREM}
\title{Simulated data for a PCREM}
\format{
A data frame with 210 rows and 4 variables
}
\usage{
data(SimData_PCREM)
}
\description{
Simulated data for a piecewise crossed random effects model (PCREM) with 7 timepoints collected on 30 individuals. The variables are as follows:
}
\details{
\itemize{
\item id - id for each individual
\item teacherid - id for each teacher
\item time - timepoints for each individual
\item y - outcome
}
}
\keyword{datasets}
