% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEND.R
\docType{data}
\name{SimData_PREM}
\alias{SimData_PREM}
\title{Simulated data for a PREM}
\format{
A data frame with 200 rows and 3 variables.
}
\usage{
data(SimData_PREM)
}
\description{
Simulated data for a piecewise random effects model (PREM) with 10 timepoints collected on 20 individuals.
}
\details{
\itemize{
\item \code{id} ID for each individual.
\item \code{time} Timepoints for each individual.
\item \code{y} Outcome.
}
}
\keyword{datasets}
