% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_CREM.R
\name{summary.CREM}
\alias{summary.CREM}
\alias{print.summary.CREM}
\title{Summarize the results of a crossed random effects model (CREM)}
\usage{
\method{summary}{CREM}(object, ...)

\method{print}{summary.CREM}(x, ...)
}
\arguments{
\item{object}{An object of class "CREM" (returned by \code{Bayes_CREM(...)}).}

\item{...}{Additional arguments.}

\item{x}{An object of class "summary.CREM".}
}
\value{
Returns a list of key parameter estimates.
}
\description{
Provides a summary of a CREM model, as returned by \code{Bayes_CREM()}.
}
\examples{
# load fitted model results
data(results_pcrem)
# result summary
summary(results_pcrem)

}
\author{
Corissa T. Rohloff
}
