% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_PREM.R
\name{summary.PREM}
\alias{summary.PREM}
\alias{print.summary.PREM}
\title{Summarize the results of a piecewise random effects model (PREM)}
\usage{
\method{summary}{PREM}(object, ...)

\method{print}{summary.PREM}(x, ...)
}
\arguments{
\item{object}{An object of class "PREM" (returned by \code{Bayes_PREM(...)}).}

\item{...}{Additional arguments.}

\item{x}{An object of class "summary.PREM".}
}
\value{
Returns a list of key parameter estimates.
}
\description{
Provides a summary of a PREM model, as returned by \code{Bayes_PREM()}.
}
\examples{
# load fitted model results
data(results_prem)
# result summary
summary(results_prem)

}
\author{
Corissa T. Rohloff
}
