% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_PREM.R
\name{summary.PREM}
\alias{summary.PREM}
\title{Summarize the results of a piecewise random effects model (PREM)}
\usage{
\method{summary}{PREM}(object, ...)
}
\arguments{
\item{object}{An object of class "PREM" (returned by \code{Bayes_PREM(...)}).}

\item{...}{Additional arguments.}
}
\value{
Prints estimates for key parameters in the PREM. Also returns a list of these values.
}
\description{
Provides a summary of a PREM model, as returned by \code{Bayes_PREM()}.
}
\examples{
\dontrun{
# load simulated data
data(SimData_PREM)
# fit Bayes_PREM
results_prem <- Bayes_PREM(data = SimData_PREM,
                           id_var = "id",
                           time_var = "time",
                           y_var = "y")
# result summary
summary(results_prem)
}

}
\author{
Corissa T. Rohloff
}
